+++ src/components/scenes/SceneEntryFocused.jsx
@@ -2,7 +2,6 @@ import { useCurrentCursor } from '@/stores/user';
 
 import BlockEntryFocused from '../blocks/BlockEntryFocused';
 import BlockList from '../blocks/BlockList';
-import SettingsSceneShotsRenumberButton from '../shots/settings/SettingsSceneShotsRenumberButton';
 import SceneEntryHeader from './SceneEntryHeader';
 
 /**
@@ -22,13 +21,6 @@ export default function SceneEntryFocused({ className, documentId }) {
     <div className="absolute top-0 left-0 right-0 bottom-0 z-10 overflow-y-auto bg-white">
       <section className={'flex flex-col mb-10' + ' ' + className}>
         <SceneEntryHeader documentId={documentId} sceneId={sceneId} />
-        <div className="flex flex-row items-center">
-          <label className="mx-2">Options:</label>
-          <SettingsSceneShotsRenumberButton
-            documentId={documentId}
-            sceneId={sceneId}
-          />
-        </div>
         {!shotId ? (
           <BlockList documentId={documentId} sceneId={sceneId} />
         ) : (